Enum rusoto_cloudfront::CreateDistributionWithTagsError

pub enum CreateDistributionWithTagsError {
    AccessDenied(String),
    CNAMEAlreadyExists(String),
    DistributionAlreadyExists(String),
    InconsistentQuantities(String),
    InvalidArgument(String),
    InvalidDefaultRootObject(String),
    InvalidErrorCode(String),
    InvalidForwardCookies(String),
    InvalidGeoRestrictionParameter(String),
    InvalidHeadersForS3Origin(String),
    InvalidLambdaFunctionAssociation(String),
    InvalidLocationCode(String),
    InvalidMinimumProtocolVersion(String),
    InvalidOrigin(String),
    InvalidOriginAccessIdentity(String),
    InvalidOriginKeepaliveTimeout(String),
    InvalidOriginReadTimeout(String),
    InvalidProtocolSettings(String),
    InvalidQueryStringParameters(String),
    InvalidRelativePath(String),
    InvalidRequiredProtocol(String),
    InvalidResponseCode(String),
    InvalidTTLOrder(String),
    InvalidTagging(String),
    InvalidViewerCertificate(String),
    InvalidWebACLId(String),
    MissingBody(String),
    NoSuchOrigin(String),
    TooManyCacheBehaviors(String),
    TooManyCertificates(String),
    TooManyCookieNamesInWhiteList(String),
    TooManyDistributionCNAMEs(String),
    TooManyDistributions(String),
    TooManyDistributionsWithLambdaAssociations(String),
    TooManyHeadersInForwardedValues(String),
    TooManyLambdaFunctionAssociations(String),
    TooManyOriginCustomHeaders(String),
    TooManyOrigins(String),
    TooManyQueryStringParameters(String),
    TooManyTrustedSigners(String),
    TrustedSignerDoesNotExist(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by CreateDistributionWithTags

Variants
AccessDenied(String)

Access denied.

CNAMEAlreadyExists(String)DistributionAlreadyExists(String)

The caller reference you attempted to create the distribution with is associated with another distribution.

InconsistentQuantities(String)

The value of Quantity and the size of Items do not match.

InvalidArgument(String)

The argument is invalid.

InvalidDefaultRootObject(String)

The default root object file name is too big or contains an invalid character.

InvalidErrorCode(String)InvalidForwardCookies(String)

Your request contains forward cookies option which doesn't match with the expectation for the whitelisted list of cookie names. Either list of cookie names has been specified when not allowed or list of cookie names is missing when expected.

InvalidGeoRestrictionParameter(String)InvalidHeadersForS3Origin(String)InvalidLambdaFunctionAssociation(String)

The specified Lambda function association is invalid.

InvalidLocationCode(String)InvalidMinimumProtocolVersion(String)InvalidOrigin(String)

The Amazon S3 origin server specified does not refer to a valid Amazon S3 bucket.

InvalidOriginAccessIdentity(String)

The origin access identity is not valid or doesn't exist.

InvalidOriginKeepaliveTimeout(String)InvalidOriginReadTimeout(String)InvalidProtocolSettings(String)

You cannot specify SSLv3 as the minimum protocol version if you only want to support only clients that support Server Name Indication (SNI).

InvalidQueryStringParameters(String)InvalidRelativePath(String)

The relative path is too big, is not URL-encoded, or does not begin with a slash (/).

InvalidRequiredProtocol(String)

This operation requires the HTTPS protocol. Ensure that you specify the HTTPS protocol in your request, or omit the RequiredProtocols element from your distribution configuration.

InvalidResponseCode(String)InvalidTTLOrder(String)InvalidTagging(String)InvalidViewerCertificate(String)InvalidWebACLId(String)MissingBody(String)

This operation requires a body. Ensure that the body is present and the Content-Type header is set.

NoSuchOrigin(String)

No origin exists with the specified Origin Id.

TooManyCacheBehaviors(String)

You cannot create more cache behaviors for the distribution.

TooManyCertificates(String)

You cannot create anymore custom SSL/TLS certificates.

TooManyCookieNamesInWhiteList(String)

Your request contains more cookie names in the whitelist than are allowed per cache behavior.

TooManyDistributionCNAMEs(String)

Your request contains more CNAMEs than are allowed per distribution.

TooManyDistributions(String)

Processing your request would cause you to exceed the maximum number of distributions allowed.

TooManyDistributionsWithLambdaAssociations(String)

Processing your request would cause the maximum number of distributions with Lambda function associations per owner to be exceeded.

TooManyHeadersInForwardedValues(String)TooManyLambdaFunctionAssociations(String)

Your request contains more Lambda function associations than are allowed per distribution.

TooManyOriginCustomHeaders(String)TooManyOrigins(String)

You cannot create more origins for the distribution.

TooManyQueryStringParameters(String)TooManyTrustedSigners(String)

Your request contains more trusted signers than are allowed per distribution.

TrustedSignerDoesNotExist(String)

One or more of your trusted signers do not exist.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateDistributionWithTagsError

fn from_body(body: &str) -> CreateDistributionWithTagsError