Enum rusoto_cloudfront::CreateStreamingDistributionWithTagsError

pub enum CreateStreamingDistributionWithTagsError {
    AccessDenied(String),
    CNAMEAlreadyExists(String),
    InconsistentQuantities(String),
    InvalidArgument(String),
    InvalidOrigin(String),
    InvalidOriginAccessIdentity(String),
    InvalidTagging(String),
    MissingBody(String),
    StreamingDistributionAlreadyExists(String),
    TooManyStreamingDistributionCNAMEs(String),
    TooManyStreamingDistributions(String),
    TooManyTrustedSigners(String),
    TrustedSignerDoesNotExist(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by CreateStreamingDistributionWithTags

Variants
AccessDenied(String)

Access denied.

CNAMEAlreadyExists(String)InconsistentQuantities(String)

The value of Quantity and the size of Items don't match.

InvalidArgument(String)

The argument is invalid.

InvalidOrigin(String)

The Amazon S3 origin server specified does not refer to a valid Amazon S3 bucket.

InvalidOriginAccessIdentity(String)

The origin access identity is not valid or doesn't exist.

InvalidTagging(String)MissingBody(String)

This operation requires a body. Ensure that the body is present and the Content-Type header is set.

StreamingDistributionAlreadyExists(String)TooManyStreamingDistributionCNAMEs(String)TooManyStreamingDistributions(String)

Processing your request would cause you to exceed the maximum number of streaming distributions allowed.

TooManyTrustedSigners(String)

Your request contains more trusted signers than are allowed per distribution.

TrustedSignerDoesNotExist(String)

One or more of your trusted signers don't exist.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateStreamingDistributionWithTagsError

pub fn from_body(body: &str) -> CreateStreamingDistributionWithTagsError