Struct rusoto_cloudfront::CloudFrontClient

pub struct CloudFrontClient { /* fields omitted */ }

A client for the CloudFront API.

Methods

impl CloudFrontClient

pub fn new(region: Region) -> CloudFrontClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CloudFrontClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,