Struct rusoto_cloudfront::CloudFrontOriginAccessIdentityList

pub struct CloudFrontOriginAccessIdentityList {
    pub is_truncated: bool,
    pub items: Option<Vec<CloudFrontOriginAccessIdentitySummary>>,
    pub marker: String,
    pub max_items: i64,
    pub next_marker: Option<String>,
    pub quantity: i64,
}

Lists the origin access identities for CloudFront.Send a GET request to the /CloudFront API version/origin-access-identity/cloudfront resource. The response includes a CloudFrontOriginAccessIdentityList element with zero or more CloudFrontOriginAccessIdentitySummary child elements. By default, your entire list of origin access identities is returned in one single page. If the list is long, you can paginate it using the MaxItems and Marker parameters.

Fields
is_truncated: bool

A flag that indicates whether more origin access identities remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more items in the list.

items: Option<Vec<CloudFrontOriginAccessIdentitySummary>>

A complex type that contains one CloudFrontOriginAccessIdentitySummary element for each origin access identity that was created by the current AWS account.

marker: String

Use this when paginating results to indicate where to begin in your list of origin access identities. The results include identities in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last identity on that page).

max_items: i64

The maximum number of origin access identities you want in the response body.

next_marker: Option<String>

If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your origin access identities where they left off.

quantity: i64

The number of CloudFront origin access identities that were created by the current AWS account.