Struct rusoto_cloudfront::CustomOriginConfig

pub struct CustomOriginConfig {
    pub http_port: i64,
    pub https_port: i64,
    pub origin_keepalive_timeout: Option<i64>,
    pub origin_protocol_policy: String,
    pub origin_read_timeout: Option<i64>,
    pub origin_ssl_protocols: Option<OriginSslProtocols>,
}

A customer origin.

Fields
http_port: i64

The HTTP port the custom origin listens on.

https_port: i64

The HTTPS port the custom origin listens on.

origin_keepalive_timeout: Option<i64>

You can create a custom keep-alive timeout. All timeout units are in seconds. The default keep-alive timeout is 5 seconds, but you can configure custom timeout lengths using the CloudFront API. The minimum timeout length is 1 second; the maximum is 60 seconds.

If you need to increase the maximum time limit, contact the AWS Support Center.

origin_protocol_policy: String

The origin protocol policy to apply to your origin.

origin_read_timeout: Option<i64>

You can create a custom origin read timeout. All timeout units are in seconds. The default origin read timeout is 30 seconds, but you can configure custom timeout lengths using the CloudFront API. The minimum timeout length is 4 seconds; the maximum is 60 seconds.

If you need to increase the maximum time limit, contact the AWS Support Center.

origin_ssl_protocols: Option<OriginSslProtocols>

The SSL/TLS protocols that you want CloudFront to use when communicating with your origin over HTTPS.