Struct rusoto_cloudfront::DistributionSummary

pub struct DistributionSummary {
    pub arn: String,
    pub aliases: Aliases,
    pub cache_behaviors: CacheBehaviors,
    pub comment: String,
    pub custom_error_responses: CustomErrorResponses,
    pub default_cache_behavior: DefaultCacheBehavior,
    pub domain_name: String,
    pub enabled: bool,
    pub http_version: String,
    pub id: String,
    pub is_ipv6_enabled: bool,
    pub last_modified_time: String,
    pub origins: Origins,
    pub price_class: String,
    pub restrictions: Restrictions,
    pub status: String,
    pub viewer_certificate: ViewerCertificate,
    pub web_acl_id: String,
}

A summary of the information about a CloudFront distribution.

Fields
arn: String

The ARN (Amazon Resource Name) for the distribution. For example: arn:aws:cloudfront::123456789012:distribution/EDFDVBD632BHDS5, where 123456789012 is your AWS account ID.

aliases: Aliases

A complex type that contains information about CNAMEs (alternate domain names), if any, for this distribution.

cache_behaviors: CacheBehaviors

A complex type that contains zero or more CacheBehavior elements.

comment: String

The comment originally specified when this distribution was created.

custom_error_responses: CustomErrorResponses

A complex type that contains zero or more CustomErrorResponses elements.

default_cache_behavior: DefaultCacheBehavior

A complex type that describes the default cache behavior if you do not specify a CacheBehavior element or if files don't match any of the values of PathPattern in CacheBehavior elements. You must create exactly one default cache behavior.

domain_name: String

The domain name that corresponds to the distribution. For example: d604721fxaaqy9.cloudfront.net.

enabled: bool

Whether the distribution is enabled to accept user requests for content.

http_version: String

Specify the maximum HTTP version that you want viewers to use to communicate with CloudFront. The default value for new web distributions is http2. Viewers that don't support HTTP/2 will automatically use an earlier version.

id: String

The identifier for the distribution. For example: EDFDVBD632BHDS5.

is_ipv6_enabled: bool

Whether CloudFront responds to IPv6 DNS requests with an IPv6 address for your distribution.

last_modified_time: String

The date and time the distribution was last modified.

origins: Origins

A complex type that contains information about origins for this distribution.

price_class: String restrictions: Restrictions status: String

The current status of the distribution. When the status is Deployed, the distribution's information is propagated to all CloudFront edge locations.

viewer_certificate: ViewerCertificate web_acl_id: String

The Web ACL Id (if any) associated with the distribution.