Struct rusoto_cloudfront::StreamingDistribution

pub struct StreamingDistribution {
    pub arn: String,
    pub active_trusted_signers: ActiveTrustedSigners,
    pub domain_name: String,
    pub id: String,
    pub last_modified_time: Option<String>,
    pub status: String,
    pub streaming_distribution_config: StreamingDistributionConfig,
}

A streaming distribution.

Fields
arn: String active_trusted_signers: ActiveTrustedSigners

A complex type that lists the AWS accounts, if any, that you included in the TrustedSigners complex type for this distribution. These are the accounts that you want to allow to create signed URLs for private content.

The Signer complex type lists the AWS account number of the trusted signer or self if the signer is the AWS account that created the distribution. The Signer element also includes the IDs of any active CloudFront key pairs that are associated with the trusted signer's AWS account. If no KeyPairId element appears for a Signer, that signer can't create signed URLs.

For more information, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

domain_name: String

The domain name that corresponds to the streaming distribution. For example: s5c39gqb8ow64r.cloudfront.net.

id: String

The identifier for the RTMP distribution. For example: EGTXBD79EXAMPLE.

last_modified_time: Option<String>

The date and time that the distribution was last modified.

status: String

The current status of the RTMP distribution. When the status is Deployed, the distribution's information is propagated to all CloudFront edge locations.

streaming_distribution_config: StreamingDistributionConfig

The current configuration information for the RTMP distribution.