Struct rusoto_cloudfront::StreamingDistributionConfig

pub struct StreamingDistributionConfig {
    pub aliases: Option<Aliases>,
    pub caller_reference: String,
    pub comment: String,
    pub enabled: bool,
    pub logging: Option<StreamingLoggingConfig>,
    pub price_class: Option<String>,
    pub s3_origin: S3Origin,
    pub trusted_signers: TrustedSigners,
}

The RTMP distribution's configuration information.

Fields
aliases: Option<Aliases>

A complex type that contains information about CNAMEs (alternate domain names), if any, for this streaming distribution.

caller_reference: String

A unique number that ensures that the request can't be replayed. If the CallerReference is new (no matter the content of the StreamingDistributionConfig object), a new streaming distribution is created. If the CallerReference is a value that you already sent in a previous request to create a streaming distribution, and the content of the StreamingDistributionConfig is identical to the original request (ignoring white space), the response includes the same information returned to the original request. If the CallerReference is a value that you already sent in a previous request to create a streaming distribution but the content of the StreamingDistributionConfig is different from the original request, CloudFront returns a DistributionAlreadyExists error.

comment: String

Any comments you want to include about the streaming distribution.

enabled: bool

Whether the streaming distribution is enabled to accept user requests for content.

logging: Option<StreamingLoggingConfig>

A complex type that controls whether access logs are written for the streaming distribution.

price_class: Option<String>

A complex type that contains information about price class for this streaming distribution.

s3_origin: S3Origin

A complex type that contains information about the Amazon S3 bucket from which you want CloudFront to get your media files for distribution.

trusted_signers: TrustedSigners

A complex type that specifies any AWS accounts that you want to permit to create signed URLs for private content. If you want the distribution to use signed URLs, include this element; if you want the distribution to use public URLs, remove this element. For more information, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.