Struct rusoto_cloudfront::StreamingDistributionSummary

pub struct StreamingDistributionSummary {
    pub arn: String,
    pub aliases: Aliases,
    pub comment: String,
    pub domain_name: String,
    pub enabled: bool,
    pub id: String,
    pub last_modified_time: String,
    pub price_class: String,
    pub s3_origin: S3Origin,
    pub status: String,
    pub trusted_signers: TrustedSigners,
}

A summary of the information for an Amazon CloudFront streaming distribution.

Fields
arn: String

The ARN (Amazon Resource Name) for the streaming distribution. For example: arn:aws:cloudfront::123456789012:streaming-distribution/EDFDVBD632BHDS5, where 123456789012 is your AWS account ID.

aliases: Aliases

A complex type that contains information about CNAMEs (alternate domain names), if any, for this streaming distribution.

comment: String

The comment originally specified when this distribution was created.

domain_name: String

The domain name corresponding to the distribution. For example: d604721fxaaqy9.cloudfront.net.

enabled: bool

Whether the distribution is enabled to accept end user requests for content.

id: String

The identifier for the distribution. For example: EDFDVBD632BHDS5.

last_modified_time: String

The date and time the distribution was last modified.

price_class: String s3_origin: S3Origin

A complex type that contains information about the Amazon S3 bucket from which you want CloudFront to get your media files for distribution.

status: String

Indicates the current status of the distribution. When the status is Deployed, the distribution's information is fully propagated throughout the Amazon CloudFront system.

trusted_signers: TrustedSigners

A complex type that specifies the AWS accounts, if any, that you want to allow to create signed URLs for private content. If you want to require signed URLs in requests for objects in the target origin that match the PathPattern for this cache behavior, specify true for Enabled, and specify the applicable values for Quantity and Items.If you don't want to require signed URLs in requests for objects that match PathPattern, specify false for Enabled and 0 for Quantity. Omit Items. To add, change, or remove one or more trusted signers, change Enabled to true (if it's currently false), change Quantity as applicable, and specify all of the trusted signers that you want to include in the updated distribution.