Struct rusoto_cloudfront::ViewerCertificate

pub struct ViewerCertificate {
    pub acm_certificate_arn: Option<String>,
    pub cloud_front_default_certificate: Option<bool>,
    pub iam_certificate_id: Option<String>,
    pub minimum_protocol_version: Option<String>,
    pub ssl_support_method: Option<String>,
}

A complex type that specifies the following:

Which SSL/TLS certificate to use when viewers request objects using HTTPS

Whether you want CloudFront to use dedicated IP addresses or SNI when you're using alternate domain names in your object names

The minimum protocol version that you want CloudFront to use when communicating with viewers

For more information, see Using an HTTPS Connection to Access Your Objects in the Amazon Amazon CloudFront Developer Guide.

Fields
acm_certificate_arn: Option<String> cloud_front_default_certificate: Option<bool> iam_certificate_id: Option<String> minimum_protocol_version: Option<String>

Specify the minimum version of the SSL/TLS protocol that you want CloudFront to use for HTTPS connections between viewers and CloudFront: SSLv3 or TLSv1. CloudFront serves your objects only to viewers that support SSL/TLS version that you specify and later versions. The TLSv1 protocol is more secure, so we recommend that you specify SSLv3 only if your users are using browsers or devices that don't support TLSv1. Note the following:

If you specify <CloudFrontDefaultCertificate>true<CloudFrontDefaultCertificate>, the minimum SSL protocol version is TLSv1 and can't be changed.

If you're using a custom certificate (if you specify a value for ACMCertificateArn or for IAMCertificateId) and if you're using SNI (if you specify sni-only for SSLSupportMethod), you must specify TLSv1 for MinimumProtocolVersion.

ssl_support_method: Option<String>

If you specify a value for ACMCertificateArn or for IAMCertificateId, you must also specify how you want CloudFront to serve HTTPS requests: using a method that works for all clients or one that works for most clients:

vip: CloudFront uses dedicated IP addresses for your content and can respond to HTTPS requests from any viewer. However, you will incur additional monthly charges.

sni-only: CloudFront can respond to HTTPS requests from viewers that support Server Name Indication (SNI). All modern browsers support SNI, but some browsers still in use don't support SNI. If some of your users' browsers don't support SNI, we recommend that you do one of the following:

Use the vip option (dedicated IP addresses) instead of sni-only.

Use the CloudFront SSL/TLS certificate instead of a custom certificate. This requires that you use the CloudFront domain name of your distribution in the URLs for your objects, for example, https://d111111abcdef8.cloudfront.net/logo.png.

If you can control which browser your users use, upgrade the browser to one that supports SNI.

Use HTTP instead of HTTPS.

Do not specify a value for SSLSupportMethod if you specified <CloudFrontDefaultCertificate>true<CloudFrontDefaultCertificate>.

For more information, see Using Alternate Domain Names and HTTPS in the Amazon CloudFront Developer Guide.